Track window ids on border rects
More robust. Splits layout still needs work.

+++ kitty/state.h
@@ -240,7 +240,7 @@ typedef struct BorderRect {
     float left, top, right, bottom;
     struct { unsigned left, top, right, bottom; } px;
     uint32_t color;
-    int border_type;
+    long long border_type;
 } BorderRect;
 
 typedef struct BorderRects {
